api/requirements.txt
fastapi
pymongo
requests
python-dotenv
certifi
dnspython
pydantic[email]
bcrypt
